+++ src/main/kotlin/net/pwall/json/JSONSerializer.kt
@@ -2,7 +2,7 @@
  * @(#) JSONSerializer.kt
  *
  * json-kotlin Kotlin JSON Auto Serialize/deserialize
- * Copyright (c) 2019, 2020 Peter Wall
+ * Copyright (c) 2019, 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -37,6 +37,7 @@ import java.util.Date
 import java.util.Enumeration
 import java.util.stream.BaseStream
 
+import net.pwall.json.JSONKotlinException.Companion.fail
 import net.pwall.json.JSONSerializerFunctions.findToJSON
 import net.pwall.json.JSONSerializerFunctions.formatISO8601
 import net.pwall.json.JSONSerializerFunctions.isSealedSubclass
@@ -89,7 +90,7 @@ object JSONSerializer {
             objClass.findToJSON()?.let { return it.call(obj) }
         }
         catch (e: Exception) {
-            throw JSONException("Error in custom toJSON - ${objClass.simpleName}", e)
+            fail("Error in custom toJSON - ${objClass.simpleName}", e)
         }
 
         when (obj) {
@@ -178,7 +179,7 @@ object JSONSerializer {
             try {
                 val v = member.getter.call(obj)
                 if (v != null && v in references)
-                    throw JSONException("Circular reference: field ${member.name} in ${obj::class.simpleName}")
+                    fail("Circular reference: field ${member.name} in ${obj::class.simpleName}")
                 if (v != null || config.hasIncludeIfNullAnnotation(annotations) || config.includeNulls || includeAll)
                     put(name, serialize(v, config, references))
             }
@@ -186,7 +187,7 @@ object JSONSerializer {
                 throw e
             }
             catch (e: Exception) {
-                throw JSONException("Error getting property ${member.name} from ${obj::class.simpleName}", e)
+                fail("Error getting property ${member.name} from ${obj::class.simpleName}", e)
             }
             finally {
                 member.isAccessible = wasAccessible