+++ src/test/kotlin/net/pwall/json/JSONStringifyTest.kt
@@ -2,7 +2,7 @@
  * @(#) JSONStringifyTest.kt
  *
  * json-kotlin Kotlin JSON Auto Serialize/deserialize
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -639,7 +639,7 @@ class JSONStringifyTest {
         val circular2 = Circular2()
         circular1.ref = circular2
         circular2.ref = circular1
-        val exception = assertFailsWith<JSONException> {
+        val exception = assertFailsWith<JSONKotlinException> {
             JSONStringify.stringify(circular1)
         }
         expect("Circular reference: field ref in Circular2") { exception.message }